refactor: dissolve mcp-trace package into proper layers
Move trace modules out of the standalone `packages/mcp-trace/` directory
into the layers where they actually belong:

- trace-core → `packages/shared/trace/` (cross-process types & utilities)
- trace-node → `src/main/trace/` (Node.js-specific tracer)
- trace-web → `src/renderer/src/trace/core/` (browser-specific tracer,
  merged with existing trace viewer directory)
- EmitterSpanProcessor → `src/main/trace/` (depends on Node `stream`)

Update all imports from `@mcp-trace/*` to `@shared/trace`,
`@main/trace/*`, and `@renderer/trace/core` respectively.
