+++ packages/shared/trace/core/spanConvert.ts
