+++ packages/shared/trace/exporters/FuncSpanExporter.ts
