+++ packages/shared/trace/index.ts
@@ -3,6 +3,5 @@ export * from './core/traceCache'
 export * from './core/traceMethod'
 export * from './exporters/FuncSpanExporter'
 export * from './processors/CacheSpanProcessor'
-export * from './processors/EmitterSpanProcessor'
 export * from './processors/FuncSpanProcessor'
 export * from './types/config'
