+++ packages/shared/trace/processors/FuncSpanProcessor.ts
