+++ packages/shared/trace/types/config.ts
