+++ src/main/ipc.ts
@@ -17,11 +17,11 @@ import {
   validateGitBashPath
 } from '@main/utils/process'
 import { handleZoomFactor } from '@main/utils/zoom'
-import type { SpanEntity, TokenUsage } from '@mcp-trace/trace-core'
 import type { UpgradeChannel } from '@shared/config/constant'
 import { MIN_WINDOW_HEIGHT, MIN_WINDOW_WIDTH } from '@shared/config/constant'
 import type { LocalTransferConnectPayload } from '@shared/config/types'
 import { IpcChannel } from '@shared/IpcChannel'
+import type { SpanEntity, TokenUsage } from '@shared/trace'
 import { extractPdfText } from '@shared/utils/pdf'
 import type {
   AgentPersistedMessage,
