+++ src/main/services/SpanCacheService.ts
@@ -1,9 +1,9 @@
 import { loggerService } from '@logger'
-import type { Attributes, SpanEntity, TokenUsage, TraceCache } from '@mcp-trace/trace-core'
-import { convertSpanToSpanEntity } from '@mcp-trace/trace-core'
 import { SpanStatusCode } from '@opentelemetry/api'
 import type { ReadableSpan } from '@opentelemetry/sdk-trace-base'
 import { HOME_CHERRY_DIR } from '@shared/config/constant'
+import type { Attributes, SpanEntity, TokenUsage, TraceCache } from '@shared/trace'
+import { convertSpanToSpanEntity } from '@shared/trace'
 import fs from 'fs/promises'
 import * as os from 'os'
 import * as path from 'path'
