+++ src/main/trace/nodeTracer.ts
@@ -6,9 +6,8 @@ import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http'
 import type { SpanProcessor } from '@opentelemetry/sdk-trace-base'
 import { BatchSpanProcessor, ConsoleSpanExporter } from '@opentelemetry/sdk-trace-base'
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node'
-
-import type { TraceConfig } from '../trace-core/types/config'
-import { defaultConfig } from '../trace-core/types/config'
+import type { TraceConfig } from '@shared/trace'
+import { defaultConfig } from '@shared/trace'
 
 export class NodeTracer {
   private static provider: NodeTracerProvider
