+++ src/renderer/src/aiCore/trace/AiSdkSpanAdapter.ts
@@ -6,9 +6,9 @@
  */
 
 import { loggerService } from '@logger'
-import type { SpanEntity, TokenUsage } from '@mcp-trace/trace-core'
 import type { Span } from '@opentelemetry/api'
 import { SpanKind, SpanStatusCode } from '@opentelemetry/api'
+import type { SpanEntity, TokenUsage } from '@shared/trace'
 
 const logger = loggerService.withContext('AiSdkSpanAdapter')
