+++ src/renderer/src/trace/core/index.ts
