+++ src/renderer/src/trace/dataHandler/MessageStreamHandler.ts
@@ -1,7 +1,7 @@
 import type { Message, MessageStream } from '@anthropic-ai/sdk/resources/messages/messages'
-import type { TokenUsage } from '@mcp-trace/trace-core'
 import type { Span } from '@opentelemetry/api'
 import { endSpan } from '@renderer/services/SpanManagerService'
+import type { TokenUsage } from '@shared/trace'
 
 export class MessageStreamHandler {
   private span: Span
